Adding new package trace-cmd v2.9 for photon
trace-cmd build requires certain libraries during
runtime that it places in the /usr/lib/traceevent/
plugins directory on installation.

perf also uses the same libraries from the same
directory. As a result, this caused conflicts during
installation of the second package if either trace-cmd
or linux-tools (that includes perf) was already
installed.

To address the above problem, the shared libraries are
generated as a subpackage of trace-cmd called
traceevent-plugins. BOth linux-tools and trace-cmd
require traceevent-plugins to be able to deparse
certain events correctly. Steven Rostedt, the author
and maintainer of trace-cmd has confirmed that the
libraries for both are kept in sync and that it is
safe to share the libraries between the two.

For dev branch, the current checking only installs
the trace-cmd binaries. It does not install the python
plugins and man pages of trace-cmd since the trace-cmd
source needs to be patched to be able to use python3
and asciidoc3.

Change-Id: I080013489038a4d748a7e45fe306c658b8ea69ba